requirements.txt
astroid==1.6.1
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
cssselect==1.0.3
data-importer==3.0.2
dicttoxml==1.7.4
Django==1.11.13
django-extra-views==0.10.0
django-haystack==2.8.1
django-import-data==0.4.2
django-oscar==1.5.4
django-phonenumber-field==1.3.0
django-tables2==1.16.0
django-treebeard==4.3
django-widget-tweaks==1.4.2
et-xmlfile==1.0.1
factory-boy==2.11.1
fake-useragent==0.1.10
Faker==0.8.15
idna==2.6
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
lxml==4.2.1
mccabe==0.6.1
mock==2.0.0
openpyxl==2.4.0
parse==1.8.2
pbr==4.0.3
phonenumberslite==8.9.5
Pillow==5.1.0
purl==1.4
pyfiglet==0.7.5
pyquery==1.4.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
requests-xml==0.2.3
six==1.10.0
sorl-thumbnail==12.4.1
stevedore==1.28.0
termcolor==1.1.0
text-unidecode==1.2
Unidecode==0.4.21
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
virtualenvwrapper-win==1.2.5
w3lib==1.19.0
wrapt==1.10.11
xlrd==1.0.0
xmljson==0.1.9
xmltodict==0.11.0
